Make score an Interactable (#1005)
* change pickup base class to interactable

* fixing imports

* creating interactable doc

* finish up doc

* bump version++

* grammar fixes

* rename cell.pickup refs to cell.interactable

* mark Interactable as non-public

* making pickups their own package

* Merge branch 'development' into rename_pickups

* bump version

* change pickup applier to effect applier

* remove _apply_score as not needed anymore

* make deletion of interactable optional

* create score location interactable

* create score location interactable

* fix tests

* rename test folder to make more sense

* added more score location tests

* bump version

* set pickup updater back to getting only pickup cells

* Merge branch 'development' into make_score_interactable

* fix tests (add serialize method for all interactables to init)

* remove score locations (Not needed).

* update score location json test

* make world map on worker more readable

* review changes

* review changes++

* grammar fix

+++ aimmo-game/simulation/interactables/interactable.py
@@ -13,6 +13,8 @@ class _Interactable(object):
 
     def __init__(self, cell):
         self.cell = cell
+        self.delete_after_effects_applied = False
+
         self.conditions = []
         self.effects = []
         self.targets = []
@@ -40,7 +42,8 @@ class _Interactable(object):
             for target in self.targets:
                 effect(target)
 
-        self.delete()
+        if self.delete_after_effects_applied:
+            self.delete()
 
     @abstractmethod
     def serialize(self):
